+++ src/mako/utility/Str.php
@@ -189,18 +189,6 @@ class Str
 	}
 
 	/**
-	 * Converts camel case to underscored.
-	 *
-	 * @deprecated
-	 * @param  string $string The input string
-	 * @return string
-	 */
-	public static function camel2underscored(string $string): string
-	{
-		return static::camelToSnake($string);
-	}
-
-	/**
 	 * Converts underscored to camel case.
 	 *
 	 * @param  string $string The input string
@@ -210,19 +198,6 @@ class Str
 	public static function snakeToCamel(string $string, bool $upper = false): string
 	{
 		return preg_replace_callback(($upper ? '/(?:^|_)(.?)/u' : '/_(.?)/u'), static fn ($matches) => mb_strtoupper($matches[1]), $string);
-	}
-
-	/**
-	 * Converts underscored to camel case.
-	 *
-	 * @deprecated
-	 * @param  string $string The input string
-	 * @param  bool   $upper  Return upper case camelCase?
-	 * @return string
-	 */
-	public static function underscored2camel(string $string, bool $upper = false): string
-	{
-		return static::snakeToCamel($string, $upper);
 	}
 
 	/**